Reset dep install. Discover strange bundle dep bug
The `tap` package bundles `import-jsx`, which in turn bundles
`@babel/core` and a bunch of other stuff.

However, even though `@babel/core` is in the bundle, it's getting placed
in the tree at the root.  This is a mistake.

Because the node is placed, and not given the `extraneous` flag, in the
ideal tree building process, when reified, it goes ahead and puts it in
place.  When reading the actual tree, we see a valid hidden lockfile,
which contains the module, and it doens't have the extraneous flag.  So
everything seems fine.

Remove the hidden lockfile, and the problem presents itself.

Corrected here through some arduous and painstaking manual installs, so
at least we have the tree in the state it *should* be for now.  But this
is a bug in arborist that must be fixed asap.

Re: https://github.com/npm/cli/issues/1597